# public/api-docs/models/openapi.yaml
openapi: 3.1.1
info:
  title: '🏛️ MiW REST api'
  description: '[UPM] MiW REST api'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  version: 2.1.4
  contact:
    name: Máster Universitario en Ingeniería Web
    url: https://miw.etsisi.upm.es/
    email: miw.etsisi@upm.es
servers:
  - url: '{schema}://{servername}:{port}/api/v1'
    description: 'MiW Production server (uses live data)'
    variables:
      schema:
        enum:
          - http
          - https
        default: https
      servername:
        default: 127.0.0.1
      port:
        default: '8000'
...
paths:
...
  /results:
    get:
      tags: [ '📝 Results' ]
      summary: Returns all results for current user
      description: 'Returns all results from the system that the user has access to.'
      operationId: api_results_cget
      security:
        - MiWApiSecurity: []
      responses:
        '200':
          description: 'Array of results'
          content:
            application/json:
              schema:
                type: object
                properties:
                  users: { description: 'Results array', type: array, items: { $ref: '#/components/schemas/Result' } }
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    post:
      tags: [ '📝 Results' ]
      summary: Creates a new result
      description: 'Creates a Result user'
      operationId: api_result_post
      security:
        - MiWApiSecurity: []
      requestBody:
        $ref: '#/components/requestBodies/ResultRequestBody'
      responses:
        '201':
          description: '`Created`: result created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          headers:
            Location:
              $ref: '#/components/headers/Location'
        '400':
          $ref: '#/components/responses/400_BadRequest_Response'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '422':
          $ref: '#/components/responses/422_UnprocessableEntity_Response'
    options:
      tags: [ '📝 Results' ]
      summary: Provides the list of HTTP supported methods.
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_result_coptions
      responses:
        '204':
          description: '`Allow` header _&lt;Response body is empty&gt;_'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'              
  '/results/{resultId}':
    get:
      tags: [ '📝 Results' ]
      summary: Retrieves a Result resource based on a single ID.
      description: 'Returns the result identified by `resultId`.'
      operationId: api_results_get
      parameters:
        - $ref: '#/components/parameters/resultId'
      security:
        - MiWApiSecurity: [ ]
      responses:
        '200':
          description: Result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    put:
      tags: [ '📝 Results' ]
      summary: Updates a result.
      description: 'Updates the result identified by `resultId`.'
      operationId: api_results_put
      parameters:
        - $ref: '#/components/parameters/resultId'
        - $ref: '#/components/parameters/ifMatchParam'
      security:
        - MiWApiSecurity: [ ]
      requestBody:
        $ref: '#/components/requestBodies/ResultRequestBody'
      responses:
        '209':
          description: '`Content Returned`: result previously existed and is now updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Result'
        '400':
          $ref: '#/components/responses/400_BadRequest_Response'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
        '412':
          $ref: '#/components/responses/412_Precondition_Failed_Response'
    delete:
      tags: [ '📝 Results' ]
      summary: Delete a result.
      description: 'Deletes the result identified by `resultId`.'
      operationId: api_results_delete
      parameters:
        - $ref: '#/components/parameters/resultId'
      security:
        - MiWApiSecurity: [ ]
      responses:
        '204':
          description: 'Result deleted &lt;Response body is empty&gt;'
        '401':
          $ref: '#/components/responses/401_Unauthorized_Response'
        '403':
          $ref: '#/components/responses/403_Forbidden_Response'
        '404':
          $ref: '#/components/responses/404_Resource_Not_Found_Response'
    options:
      tags: [ '📝 Results' ]
      summary: Provides the list of HTTP supported methods.
      description: 'Return a `Allow` header with a comma separated list of HTTP supported methods.'
      operationId: api_results_options_id
      parameters:
        - $ref: '#/components/parameters/resultId'
      responses:
        '204':
          description: '`Allow` header _&lt;Response body is empty&gt;_'
          headers:
            Allow:
              $ref: '#/components/headers/Allow'
...
components:
  headers:
    ETag:
      description: Entity Tag (unique identifier provided for cache validation)
      schema: { type: string, pattern: "[0-9a-f]*" }
    Location:
      description: "`URI` of the new resource"
      schema: { type: string, format: url }
    Allow:
      description: List of HTTP supported methods
      schema: { type: string }
  schemas:
    User:
      type: object
      properties:
        id:
          description: 'User Id'
          type: integer
          format: int64
        email:
          description: 'User email'
          type: string
          format: email
        roles:
          description: 'User roles'
          type: string
          format: json
        _links:
          description: link relations
          type: object
          properties:
            parent:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
            self:
              type: object
              properties: {"href": { "type": "string", "format": "url" }}
      example:
        user:
          id: 1508
          email: 'miw.etsisi@upm.es'
          roles: [ 'ROLE_USER' ]
          _links:
            parent:
              href: /api/v1/users
            self:
              href: /api/v1/users/1508
    Result:
      type: object
      properties:
        id:
          description: 'Result Id'
          type: integer
          format: int64
        result:
          description: 'Result value'
          type: integer
          format: int64
        user:
          description: 'User who score the result'
          $ref: '#/components/schemas/User'
        time:
          description: 'Result date in Y-m-d H:i:s format'
          type: datetime
          format: 'Y-m-d H:i:s'
      example:
        result:
          id: 1
          result: 100
          user:
            id: 5
            email: 'user5@alumnos.upm.es'
            roles: [ 'ROLE_USER' ]
            _links:
              parent:
                href: /api/v1/users
              self:
                href: /api/v1/users/1508
          time: '2023-12-18 20:20:20'
    Message:
      type: object
      description: HTTP response message
      properties:
        code:
          description: 'Response code'
          type: integer
          format: int64
        message:
          description: 'Response message'
          type: string
      example:
        code: 200
        message: Ok
  responses:
    400_BadRequest_Response:
      description: "`BAD REQUEST`: e-mail already exists"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-400:
              $ref: '#/components/examples/response-400'
    401_Unauthorized_Response:
      description: '`Unauthorized`: invalid Authorization header'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-401:
              $ref: '#/components/examples/response-401'
    403_Forbidden_Response:
      description: '`Forbidden`: you don''t have permission to access'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-403:
              $ref: '#/components/examples/response-403'
    404_Resource_Not_Found_Response:
      description: '`Not found`: resource not found'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-404:
              $ref: '#/components/examples/response-404'
    412_Precondition_Failed_Response:
      description: '`PRECONDITION FAILED`: one or more conditions given evaluated to false'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-412:
              $ref: '#/components/examples/response-412'
    422_UnprocessableEntity_Response:
      description: "`UNPROCESSABLE ENTITY`: e-mail or password is left out"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Message'
          examples:
            response-422:
              $ref: '#/components/examples/response-422'
  requestBodies:
...
    ResultRequestBody:
      description: '`Result` data'
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              result:
                description: 'Result scored'
                type: integer
                format: int64
              user:
                description: 'User email'
                type: string
                format: email
              time:
                description: 'Result time'
                type: datetime
                format: 'Y-m-d H:i:s'
            example:
              result: 1
              user: 'user1@alumnos.upm.es'
              time: '2023-12-18 20:20:20'
  parameters:
...
    resultId:
      name: resultId
      in: path
      description: 'ID of result'
      required: true
      schema: { "type": "integer", pattern: '^\d+$' }
    ifMatchParam:
      name: If-Match
      in: header
      description: "ETag value of the resource you are trying to update"
      required: true
      schema:
        type: string
        pattern: "[0-9a-f]*"
  examples:
    response-400:
      value:
        code: 400
        message: 'BAD REQUEST: e-mail already exists'
    response-401:
      value:
        code: 401
        message: 'UNAUTHORIZED: invalid Authorization header'
    response-403:
      value:
        code: 403
        message: "FORBIDDEN You don't have permission to access"
    response-404:
      value:
        code: 404
        message: 'NOT FOUND: Resource not found'
    response-412:
      value:
        code: 412
        message: 'PRECONDITION FAILED: one or more conditions given evaluated to false'
    response-422:
      value:
        code: 422
        message: 'UNPROCESSABLE ENTITY: e-mail or password is left out'
  securitySchemes:
    MiWApiSecurity:
      type: http
      bearerFormat: JWT
      scheme: bearer